requirements.txt
streamlit
google-cloud-firestore
google-generativeai
firebase-admin